+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- [client] Allow the creation of a `Display` from a foreign `wl_display`
+
 # 0.20.8 - 2018-05-22
 
 - [client/server] Use `Clone` trait rather than inherent method for `clone()`.